Handle deep link event in POCustomTabAuthorizationActivity
